# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "June API",
    "description": "HTTP boundary for the June brain. The canonical client is apps/web (SvelteKit) but any shell speaking JSON/SSE can use it.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/skills/{key}/writes/{user_id}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "List Skill Writes",
        "description": "List paraphrased facts this skill has written for a user.\n\nFilters the vector store's shadow table by ``source LIKE 'skill:<key>:%'``,\nso it picks up writes from every tool the skill exposes. Each record\ncarries the prefixed memory ref so the UI can deep-link into ``/memory``\nor hand the ref to ``forget``.",
        "operationId": "list_skill_writes_skills__key__writes__user_id__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillWritesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/skills/{key}/tools/{tool}/toggle": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Toggle Skill Tool",
        "description": "Enable or disable a single tool inside a skill.\n\nThe skill subprocess keeps running and continues to advertise this\ntool, but the supervisor filters it out of the agent's bound tools\nat the next reload. Persists to ``skills.toml`` so the gate\nsurvives restart.",
        "operationId": "toggle_skill_tool_skills__key__tools__tool__toggle_post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "tool",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillToolToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillToolToggleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/skills/{key}/toggle": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Toggle Skill",
        "description": "Enable or disable a skill.\n\nPersists to ``~/Library/Application Support/June/skills.toml`` and\nrebuilds the agent so the next chat turn picks up the new tool set.",
        "operationId": "toggle_skill_skills__key__toggle_post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillToggleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/skills/registry/{key}/install": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Install Registry Entry",
        "description": "Materialize a registry entry into the user's skill manifest.",
        "operationId": "install_registry_entry_skills_registry__key__install_post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryInstallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/skills/registry/{key}": {
      "delete": {
        "tags": [
          "skills"
        ],
        "summary": "Uninstall Registry Entry",
        "description": "Remove an installed registry skill from the manifest.",
        "operationId": "uninstall_registry_entry_skills_registry__key__delete",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryUninstallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegistryInstallResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "installed": {
            "type": "boolean",
            "title": "Installed"
          },
          "requires_env": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requires Env"
          }
        },
        "type": "object",
        "required": [
          "key",
          "installed"
        ],
        "title": "RegistryInstallResponse",
        "description": "POST /skills/registry/{key}/install payload."
      },
...
      "RegistryUninstallResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "uninstalled": {
            "type": "boolean",
            "title": "Uninstalled"
          }
        },
        "type": "object",
        "required": [
          "key",
          "uninstalled"
        ],
        "title": "RegistryUninstallResponse",
        "description": "DELETE /skills/registry/{key} payload."
      },
...
      "SkillToggleRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "New enabled state."
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "SkillToggleRequest",
        "description": "POST /skills/{key}/toggle body."
      },
      "SkillToggleResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "key",
          "enabled",
          "status"
        ],
        "title": "SkillToggleResponse",
        "description": "Result of a toggle \u2014 callers refresh their UI from this."
      },
...
      "SkillToolToggleRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "New enabled state for this tool."
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "SkillToolToggleRequest",
        "description": "POST /skills/{key}/tools/{tool}/toggle body."
      },
      "SkillToolToggleResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "key",
          "tool",
          "enabled"
        ],
        "title": "SkillToolToggleResponse",
        "description": "Result of a per-tool toggle."
      },
      "SkillWriteRecord": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref",
            "description": "Memory ref, e.g. \"semantic:abc...\"."
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Paraphrased fact text the recall pipeline sees."
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Full source tag, e.g. \"skill:daily:track_goal\"."
          },
          "tool": {
            "type": "string",
            "title": "Tool",
            "description": "Tool name extracted from the source tag.",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp when the write landed.",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "ref",
          "text",
          "source"
        ],
        "title": "SkillWriteRecord",
        "description": "One paraphrased fact a skill has written into the vector store."
      },
      "SkillWritesResponse": {
        "properties": {
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "writes": {
            "items": {
              "$ref": "#/components/schemas/SkillWriteRecord"
            },
            "type": "array",
            "title": "Writes"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "skill",
          "user_id"
        ],
        "title": "SkillWritesResponse",
        "description": "GET /skills/{key}/writes payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}